requirements.txt
streamlit>=1.28.0
google-generativeai>=0.8.0
supabase>=2.0.0
requests>=2.28.0
pandas>=2.0.0
openpyxl>=3.1.0
python-docx>=1.0.0
